@@ -6,11 +6,11 @@ ruby "2.6.5"
 gem "dotenv-rails", groups: [:development, :test]
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem "rails", "~> 5.2.6"
+gem "rails", "~> 6.0.4.1"
 gem "rake"
 
 # JSON serializer
-gem "active_model_serializers", "0.9.3"
+gem "active_model_serializers", "0.9.8"
 
 # Use postgresql and mongo as the database for Active Record
 gem "pg"
@@ -21,7 +21,7 @@ gem "puma"
 
 # Authentication libraries
 gem "devise"
-gem "cancancan", "~> 2.2.0"
+gem "cancancan", "~> 3.0.0"
 gem "cancancan-mongoid"
 gem "devise_invitable"
 gem "omniauth", "~> 1.8.1"
@@ -86,11 +86,7 @@ group :development do
   gem "better_errors", "~>2.1"
   gem "binding_of_caller", "~>0.7"
   gem "foreman", require: false
-  # thor > 0.19.1 prints some warnings to stderr
-  # This is fixed is thor > 0.20.0 but 0.19 is required by foreman.
-  # Foreman suggests not including foreman in your Gemfile
-  # https://github.com/erikhuda/thor/issues/538
-  gem "thor", "0.19.1"
+  gem "thor", "~> 0.20"
   gem "letter_opener"
   gem "rubocop-rails", require: false
 end
