@@ -6,7 +6,7 @@ class Api::V1::UsersController < ApplicationController
   end
 
   def update
-    @user.update_attributes!(user_params)
+    @user.update!(user_params)
 
     render json: @user
   end
