@@ -7,7 +7,7 @@ class MergeTrackables::TopicFollowing
     TopicFollowing.all(trackable_key.to_sym.in => rest_ids).map do |tf|
       updated_ids = tf.send(trackable_key) - rest_ids + [parent_id]
 
-      tf.update_attributes(trackable_key.to_sym => updated_ids.uniq)
+      tf.update(trackable_key.to_sym => updated_ids.uniq)
     end
 
     p "PERFORM MergeTrackables::CheckinTrackables"
