@@ -5,6 +5,6 @@ class PositionReferenceJob
     object = object_class.constantize.find_by(id: objec_id)
     position = Position.find_or_create_by(postal_code: postal_code)
 
-    object.update_attributes(position_id: position.id) if position.persisted?
+    object.update(position_id: position.id) if position.persisted?
   end
 end
